Make DictEncodedObservations be a detail of InitialEnsembleRunModel
Avoids sharing this type with model_factory and storage
