+++ src/ert/run_models/model_factory.py
@@ -6,7 +6,6 @@ from queue import SimpleQueue
 from typing import TYPE_CHECKING
 
 import numpy as np
-import polars as pl
 
 from ert.config import (
     ConfigValidationError,
@@ -25,7 +24,6 @@ from ert.mode_definitions import (
 )
 from ert.validation import ActiveRange
 
-from ..storage.local_experiment import DictEncodedObservations
 from .ensemble_experiment import EnsembleExperiment
 from .ensemble_information_filter import EnsembleInformationFilter
 from .ensemble_smoother import EnsembleSmoother
@@ -43,14 +41,6 @@ if TYPE_CHECKING:
 
 
 logger = logging.getLogger(__name__)
-
-
-def serialize_observations(
-    observations: dict[str, pl.DataFrame],
-) -> dict[str, DictEncodedObservations]:
-    return {
-        k: DictEncodedObservations.from_polars(df) for k, df in observations.items()
-    }
 
 
 def create_model(
@@ -122,7 +112,7 @@ def _setup_single_test_run(
         log_path=config.analysis_config.log_path,
         storage_path=config.ens_path,
         queue_config=config.queue_config.create_local_copy(),
-        observations=serialize_observations(config.observations),
+        observations=config.observations,
         status_queue=status_queue,
     )
 
@@ -172,7 +162,7 @@ def _setup_ensemble_experiment(
         log_path=config.analysis_config.log_path,
         storage_path=config.ens_path,
         queue_config=config.queue_config,
-        observations=serialize_observations(config.observations),
+        observations=config.observations,
         status_queue=status_queue,
     )
 
@@ -260,7 +250,7 @@ def _setup_manual_update(
         substitutions=config.substitutions,
         hooked_workflows=config.hooked_workflows,
         log_path=config.analysis_config.log_path,
-        observations=serialize_observations(config.observations),
+        observations=config.observations,
     )
 
 
@@ -301,7 +291,7 @@ def _setup_ensemble_smoother(
         substitutions=config.substitutions,
         hooked_workflows=config.hooked_workflows,
         log_path=config.analysis_config.log_path,
-        observations=serialize_observations(config.observations),
+        observations=config.observations,
     )
 
 
@@ -342,7 +332,7 @@ def _setup_ensemble_information_filter(
         substitutions=config.substitutions,
         hooked_workflows=config.hooked_workflows,
         log_path=config.analysis_config.log_path,
-        observations=serialize_observations(config.observations),
+        observations=config.observations,
     )
 
 
@@ -405,7 +395,7 @@ def _setup_multiple_data_assimilation(
         substitutions=config.substitutions,
         hooked_workflows=config.hooked_workflows,
         log_path=config.analysis_config.log_path,
-        observations=serialize_observations(config.observations),
+        observations=config.observations,
     )
