+++ src/ert/storage/local_experiment.py
@@ -6,7 +6,7 @@ from collections.abc import Generator
 from datetime import datetime
 from functools import cached_property
 from pathlib import Path
-from typing import TYPE_CHECKING, Annotated, Any, Literal, Self
+from typing import TYPE_CHECKING, Annotated, Any
 from uuid import UUID
 
 import polars as pl
@@ -33,8 +33,6 @@ if TYPE_CHECKING:
     from .local_ensemble import LocalEnsemble
     from .local_storage import LocalStorage
 
-from polars.datatypes import DataTypeClass
-
 
 class _Index(BaseModel):
     id: UUID
@@ -57,37 +55,6 @@ _parameters_adapter = TypeAdapter(  # type: ignore
         Field(discriminator="type"),
     ]
 )
-
-
-# https://github.com/pola-rs/polars/issues/13152#issuecomment-1864600078
-# PS: Serializing/deserializing schema is scheduled to be added to polars core,
-# ref https://github.com/pola-rs/polars/issues/20426
-# then this workaround can be omitted.
-def str_to_dtype(dtype_str: str) -> pl.DataType:
-    dtype = eval(f"pl.{dtype_str}")
-    if isinstance(dtype, DataTypeClass):
-        dtype = dtype()
-    return dtype
-
-
-class DictEncodedObservations(BaseModel):
-    type: Literal["dicts"]
-    data: list[dict[str, Any]]
-    datatypes: dict[str, str]
-
-    @classmethod
-    def from_polars(cls, data: pl.DataFrame) -> Self:
-        str_schema = {k: str(dtype) for k, dtype in data.schema.items()}
-        return cls(type="dicts", data=data.to_dicts(), datatypes=str_schema)
-
-    def to_polars(self) -> pl.DataFrame:
-        return pl.from_dicts(
-            self.data,
-            schema={
-                col: str_to_dtype(dtype_str)
-                for col, dtype_str in self.datatypes.items()
-            },
-        )
 
 
 class LocalExperiment(BaseMode):
@@ -138,7 +105,7 @@ class LocalExperiment(BaseMode):
         *,
         parameters: list[ParameterConfig] | None = None,
         responses: list[ResponseConfig] | None = None,
-        observations: dict[str, DictEncodedObservations] | None = None,
+        observations: dict[str, pl.DataFrame] | None = None,
         simulation_arguments: dict[Any, Any] | None = None,
         name: str | None = None,
         templates: list[tuple[str, str]] | None = None,
@@ -220,7 +187,7 @@ class LocalExperiment(BaseMode):
             output_path.mkdir()
             for response_type, dataset in observations.items():
                 storage._to_parquet_transaction(
-                    output_path / f"{response_type}", dataset.to_polars()
+                    output_path / f"{response_type}", dataset
                 )
 
         simulation_data = simulation_arguments or {}
